% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-world-wtmp.R
\name{app_world_wtmp}
\alias{app_world_wtmp}
\title{World Sea Temp shiny app}
\usage{
app_world_wtmp(wtmp_data_file, zoom_locs_file)
}
\arguments{
\item{zoom_locs_file}{}
}
\value{

}
\description{
World Sea Temp shiny app
}
