% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station_info.R
\name{avail_buoy_years}
\alias{avail_buoy_years}
\title{Find available buoy years}
\usage{
avail_buoy_years()
}
\value{
a tibble with two columns: "buoy" and "year"
}
\description{
For each NDBC buoy, which years are available?
}
