% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{avail_years_for_buoy}
\alias{avail_years_for_buoy}
\title{Find available years for a single buoy}
\usage{
avail_years_for_buoy(id)
}
\value{
vector of years
}
\description{
Find available years for a single buoy
}
