% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tahoe.R
\name{getAvg}
\alias{getAvg}
\title{Take the mean from a min-mean-max string, Lake Tahoe JPL data}
\usage{
getAvg(minavgmax)
}
\arguments{
\item{minavgmax}{}
}
\value{

}
\description{
e.g., "15:15.2:15.6" --> 15.2
}
