% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mult.R
\name{insist_read_months}
\alias{insist_read_months}
\title{Read and bind multiple monthly datasets}
\usage{
insist_read_months(...)
}
\arguments{
\item{buoy_id}{}
}
\value{

}
\description{
Read and bind multiple monthly datasets
}
