% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mult.R
\name{insist_read_years}
\alias{insist_read_years}
\title{Read and bind multiple yearly datasets}
\usage{
insist_read_years(...)
}
\arguments{
\item{years}{list of years}
}
\value{

}
\description{
Read and bind multiple yearly datasets
}
