% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ndbc_45d}
\alias{ndbc_45d}
\title{Get last 45 days' data from a NDBC buoy}
\usage{
ndbc_45d(buoy_id)
}
\arguments{
\item{buoy_id}{}
}
\value{
a data frame
}
\description{
Get last 45 days' data from a NDBC buoy
}
