% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ndbc_5d}
\alias{ndbc_5d}
\title{Get last 5 days' data from a NDBC buoy}
\usage{
ndbc_5d(buoy_id)
}
\arguments{
\item{buoy_id}{}
}
\value{
a data frame
}
\description{
Get last 5 days' data from a NDBC buoy
}
