% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{ndbc_agg_by_doy}
\alias{ndbc_agg_by_doy}
\title{Aggregate daily data by day of year}
\usage{
ndbc_agg_by_doy(daily_data, fn = c(mean, min, max))
}
\arguments{
\item{fn}{}
}
\value{

}
\description{
Aggregate daily data by day of year
}
