% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mult.R
\name{ndbc_all_hist}
\alias{ndbc_all_hist}
\title{Download, merge, & munge all historical data-years for a NDBC buoy}
\usage{
ndbc_all_hist(buoy_id)
}
\arguments{
\item{buoy_id}{}
}
\value{

}
\description{
Download, merge, & munge all historical data-years for a NDBC buoy
}
