% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ndbc_histM}
\alias{ndbc_histM}
\title{Get historical (quality-controlled) data from a NDBC buoy,
for a given month during the past year.}
\usage{
ndbc_histM(buoy_id, month)
}
\arguments{
\item{month}{}
}
\value{
a data frame
}
\description{
Get historical (quality-controlled) data from a NDBC buoy,
for a given month during the past year.
}
