% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ndbc_histY}
\alias{ndbc_histY}
\title{Get historical (quality-controlled) data from a NDBC buoy for a given year.}
\usage{
ndbc_histY(buoy_id, year)
}
\arguments{
\item{year}{}
}
\value{
a data frame
}
\description{
Get historical (quality-controlled) data from a NDBC buoy for a given year.
}
