% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdmet_hourly.R
\name{ndbc_hourly_all}
\alias{ndbc_hourly_all}
\title{Read all hours of std-met data from NDBC 'hourly2' directory}
\usage{
ndbc_hourly_all()
}
\value{

}
\description{
Read all hours of std-met data from NDBC 'hourly2' directory
}
