% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdmet_hourly.R
\name{ndbc_hourly_hour}
\alias{ndbc_hourly_hour}
\title{Read one hour of std-met data from NDBC 'hourly2' directory}
\usage{
ndbc_hourly_hour(hour)
}
\arguments{
\item{hour}{}
}
\value{

}
\description{
Read one hour of std-met data from NDBC 'hourly2' directory
}
