% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdmet_hourly.R
\name{ndbc_hourly_hour_base}
\alias{ndbc_hourly_hour_base}
\title{Read one hour of std-met data from NDBC 'hourly2' directory (Base R)}
\usage{
ndbc_hourly_hour_base(hour)
}
\value{
data.frame
}
\description{
Read one hour of std-met data from NDBC 'hourly2' directory (Base R)
}
