% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdmet_hourly.R
\name{ndbc_hourly_some}
\alias{ndbc_hourly_some}
\title{Read some hours of std-met data from NDBC 'hourly2' directory}
\usage{
ndbc_hourly_some(hours)
}
\arguments{
\item{hours}{vector of integers (range 0-23)}
}
\value{

}
\description{
Read some hours of std-met data from NDBC 'hourly2' directory
}
