% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ndbc_read}
\alias{ndbc_read}
\title{Download and read a plain-text data file from NDBC}
\usage{
ndbc_read(url)
}
\value{
a data frame
}
\description{
Download and read a plain-text data file from NDBC
}
\examples{
ndbc_read('https://www.ndbc.noaa.gov/data/5day2/46053_5day.txt')
ndbc_read('https://www.ndbc.noaa.gov/data/historical/stdmet/46053h2019.txt.gz')
}
