% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ndbc_read_45day}
\alias{ndbc_read_45day}
\title{Get last 45 days' data from a NDBC buoy}
\usage{
ndbc_read_45day(buoy_id)
}
\arguments{
\item{buoy_id}{}
}
\value{
a data frame
}
\description{
Get last 45 days' data from a NDBC buoy
}
