% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ndbc_read_cols}
\alias{ndbc_read_cols}
\title{Read column names from a NDBC std-met data}
\usage{
ndbc_read_cols(url)
}
\arguments{
\item{url}{}
}
\value{
vector of names
}
\description{
Read column names from a NDBC std-met data
}
\examples{
ndbc_read_cols('https://www.ndbc.noaa.gov/data/5day2/46053_5day.txt')
}
