% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{ndbc_read_month_recent}
\alias{ndbc_read_month_recent}
\title{Get historical (quality-controlled) data from a NDBC buoy,
for a given month during the past year.}
\usage{
ndbc_read_month_recent(buoy_id, month)
}
\arguments{
\item{month}{(1-12)}
}
\value{
a data frame
}
\description{
Get historical (quality-controlled) data from a NDBC buoy,
for a given month during the past year.
}
