% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mult.R
\name{ndbc_read_mult_month}
\alias{ndbc_read_mult_month}
\title{Read and bind multiple monthly datasets}
\usage{
ndbc_read_mult_month(buoy_id, start_month, end_month)
}
\arguments{
\item{buoy_id}{}
}
\value{

}
\description{
Read and bind multiple monthly datasets
}
