% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mult.R
\name{ndbc_read_mult_year}
\alias{ndbc_read_mult_year}
\title{Read and bind multiple yearly datasets}
\usage{
ndbc_read_mult_year(buoy_id, start_year, end_year)
}
\arguments{
\item{buoy_id}{}
}
\value{

}
\description{
Read and bind multiple yearly datasets
}
