% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ndbc_read_txt}
\alias{ndbc_read_txt}
\title{Download and read a plain-text data file from NDBC}
\usage{
ndbc_read_txt(url)
}
\value{
a data frame
}
\description{
Download and read a plain-text data file from NDBC
}
