% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mult.R
\name{ndbc_read_years}
\alias{ndbc_read_years}
\title{Read and bind multiple yearly datasets}
\usage{
ndbc_read_years(buoy_id, years)
}
\arguments{
\item{years}{list of years}
}
\value{

}
\description{
Read and bind multiple yearly datasets
}
