% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ndbc_read_zip}
\alias{ndbc_read_zip}
\title{Download and read a gzipped historical data file from NDBC}
\usage{
ndbc_read_zip(url)
}
\value{
a data frame
}
\description{
Download and read a gzipped historical data file from NDBC
}
