% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{ndbc_reduce_daily}
\alias{ndbc_reduce_daily}
\title{Daily means for one feature}
\usage{
ndbc_reduce_daily(data, feature)
}
\arguments{
\item{feature}{}
}
\value{

}
\description{
Daily means for one feature
}
