% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mult.R
\name{ndbc_window}
\alias{ndbc_window}
\title{Read std met data from arbitrary window of dates}
\usage{
ndbc_window(buoy_id, start_date, end_date)
}
\value{

}
\description{
Read std met data from arbitrary window of dates
}
