% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mult.R
\name{ndbc_window_current_year}
\alias{ndbc_window_current_year}
\title{Read std met data from arbitrary window of dates in current year}
\usage{
ndbc_window_current_year(buoy_id, start_date, end_date)
}
\value{

}
\description{
Read std met data from arbitrary window of dates in current year
}
