% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{ndbc_write_all_hist}
\alias{ndbc_write_all_hist}
\title{Get all historical data from a list of multiple NDBC buoys,
write each to a local file.}
\usage{
ndbc_write_all_hist(buoys, dir = "~/Downloads")
}
\arguments{
\item{buoys}{list of buoy IDs}

\item{dir}{local directory to write files}
}
\value{

}
\description{
Get all historical data from a list of multiple NDBC buoys,
write each to a local file.
}
