% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{recodeFeatures}
\alias{recodeFeatures}
\title{Friendlier column names for NDBC std-met data}
\usage{
recodeFeatures(data)
}
\arguments{
\item{data}{data-frame}
}
\value{

}
\description{
Friendlier column names for NDBC std-met data
}
