% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munge.R
\name{reduceToDaily}
\alias{reduceToDaily}
\title{Reduce complete buoy data to one observation per day}
\usage{
reduceToDaily(buoy_data, feature)
}
\arguments{
\item{buoy_data}{complete data file for a buoy (.csv.gz)}

\item{feature}{string (WTMP, ATMP, WSPD)}
}
\value{
data.frame - Daily means for one feature
}
\description{
Reduce complete buoy data to one observation per day
}
\examples{
reduceToDaily("FTPC1", "WTMP")
}
