% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-snapshot.R
\name{snapshotApp}
\alias{snapshotApp}
\title{Shiny app to plot snapshot of NDBC marine weather data}
\usage{
snapshotApp()
}
\value{

}
\description{
Shiny app to plot snapshot of NDBC marine weather data
}
