% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{avail_buoy_years}
\alias{avail_buoy_years}
\title{Find available historical year data for every NDBC buoy}
\usage{
avail_buoy_years()
}
\value{
a tibble with two columns: "buoy" and "year"
}
\description{
Find available historical year data for every NDBC buoy
}
