% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{getAllHist}
\alias{getAllHist}
\title{Download, merge, & munge all historical data-years for a NDBC buoy}
\usage{
getAllHist(id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Download, merge, & munge all historical data-years for a NDBC buoy
}
